merge_agent/requirements.txt
Flask==3.0.2
GitPython==3.1.41
openai==1.12.0
pandas==2.2.0
PyGithub==2.2.0
pytest==6.2.5
Requests==2.31.0
